+++ src/components/private/index.js
@@ -4,7 +4,7 @@ export const Landing = lazy(() => import("./landing/Landing"));
 export const Marquee = lazy(() => import("./marquee/Marquee"));
 export const About = lazy(() => import("./about/About"));
 export const Timeline = lazy(() => import("./timeline/Timeline"));
-export const Timer = lazy(() => import("./timer/Timer"));
+export const Timer = lazy(() => import("./timeline/Timeline"));
 export const Theme = lazy(() => import("./themes/Theme"));
 export const Prizes = lazy(() => import("./prizes/Prizes"));
 export const Sponsors = lazy(() => import("./sponsors/Sponsors"));
