+++ src/components/private/timer/Timeline.scss
@@ -1,0 +1,140 @@
+.timeline__parent {
+    padding: 0rem 0rem;
+    margin-top: 0rem;
+    position: relative;
+    overflow: hidden;
+  
+    @media screen and (max-width: 576px) {
+      padding: 3rem 1.5rem;
+    }
+  
+    .body {
+      display: flex;
+      align-items: center;
+  
+      @media screen and (max-width: 576px) {
+        justify-content: center;
+      }
+  
+      .timer {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        gap: 4rem;
+        margin-top: -4rem;
+  
+        @media screen and (max-width: 576px) {
+          margin-top: 0rem;
+          gap: 3rem;
+        }
+  
+        .header_1 {
+          font-family: "Montserrat", sans-serif;
+          letter-spacing: 2px;
+          font-size: 100px;
+          text-align: center;
+          color: white;
+          font-weight: 800;
+          line-height: 1;
+          z-index: 1;
+  
+          @media screen and (max-width: 576px) {
+            font-size: 48px;
+          }
+        }
+  
+        .countdown {
+          display: flex;
+          justify-content: center;
+          align-items: center;
+          gap: 1rem;
+          z-index: 1;
+  
+          @media screen and (max-width: 576px) {
+            gap: 10px;
+          }
+  
+          .time {
+            font-family: "Montserrat", sans-serif;
+            font-size: 60px;
+            text-align: center;
+            color: white;
+            font-weight: 800;
+            line-height: 1;
+            background-color: #2a2727;
+            padding: 1rem;
+            border-radius: 18px;
+  
+            @media screen and (max-width: 576px) {
+              font-size: 28px;
+            }
+  
+            p:nth-child(2) {
+              font-size: 18px;
+              font-weight: 600;
+  
+              @media screen and (max-width: 576px) {
+                font-size: 15px;
+              }
+            }
+          }
+  
+          .colon {
+            font-size: 60px;
+          }
+        }
+      }
+    }
+  
+    .mask {
+      right: 0;
+      top: 0;
+      position: absolute;
+      width: 40%;
+  
+      @media screen and (max-width: 576px) {
+        width: 100%;
+      }
+    }
+  
+    .letters {
+      width: 45%;
+      opacity: 0.7;
+      right: 9rem;
+      top: 6rem;
+      position: absolute;
+  
+      @media screen and (max-width: 576px) {
+        display: none;
+      }
+    }
+  
+    .ghori {
+      width: 55%;
+  
+      @media screen and (max-width: 576px) {
+        display: none;
+      }
+    }
+  
+    .text {
+      width: 35%;
+    }
+  }
+  
+  .colon {
+    animation: blink 2s infinite; // Adjust the animation duration as needed
+  }
+  
+  @keyframes blink {
+    0% {
+      opacity: 1;
+    }
+    50% {
+      opacity: 0;
+    }
+    100% {
+      opacity: 1;
+    }
+  }
