--- src/components/private/timer/Timer.scss
@@ -1,139 +1,0 @@
-.timeline {
-  margin-top: 6rem;
-  padding: 4rem 5rem 0 5rem;
-
-  @media screen and (max-width: 768px) {
-    padding: 1rem;
-  }
-  @media screen and (max-width: 344px) {
-    padding: 1rem;
-  }
-
-  .timeline__content {
-    display: flex;
-    justify-content: center;
-    align-items: flex-start;
-    flex-wrap: wrap;
-    gap: 6rem;
-    margin-top: 4rem;
-
-    @media screen and (max-width: 900px) {
-      padding: 2rem;
-      gap:4rem 9rem ;
-    }
-    @media screen and (max-width: 412px) {
-      padding: 2rem 1rem;
-      gap: 2rem 4rem ;
-    }
-    @media screen and (max-width: 540px) {
-      gap: 2rem 6rem;
-    }
-    @media screen and (max-width: 400px) {
-      gap: 2rem 4rem;
-    }
-
-
-
-    .timeline__event {
-      display: flex;
-      align-items: center;
-      flex-direction: column;
-      justify-content: center;
-
-      .timeline__d-flex {
-        display: flex;
-        align-items: center;
-        position: relative;
-        justify-content: space-between;
-
-        .timeline__number {
-          font-size: clamp(5rem, 12vw, 12rem);
-          font-weight: bolder;
-          font-family: var(--body-font);
-          position: relative;
-          margin-right: 2rem;
-
-          &.active {
-            color: #ffffff;
-          }
-
-          &.inactive {
-            color: #646464;
-          }
-        }
-       
-
-        .timeline__circle_details {
-          position: absolute;
-          border-radius: 50%;
-          width: 7rem;
-          height: 7rem;
-          text-align: center;
-          display: flex;
-          justify-content: center;
-          align-items: center;
-          left: 96%;
-          transform: translateX(-50%);
-          overflow: hidden;
-
-          @media screen and (max-width: 1024px) {
-            left: 103%;
-          }
-          @media screen and (max-width: 860px) {
-            left: 109%;
-          }
-          @media screen and (max-width: 820px) {
-            width: 6rem;
-            height: 6rem;
-            left: 100%;
-          }
-          
-          @media screen and (max-width: 540px) {
-            width: 5rem;
-            left: 108%;
-            height: 5rem;
-            left: 93%;
-          }
-          img {
-            height: 100%;
-            width: 100%;
-            display: block;
-            border-radius: 50%;
-            background-color: #1b1b1b;
-          }
-
-          p {
-            font-family: var(--body-font);
-            font-size: 1rem;
-            position: absolute;
-            color: white;
-            padding: 5px;
-
-            @media screen and (max-width: 853px) {
-              font-size: 0.7rem;
-            }
-          }
-        }
-      }
-     
-      .timeline__date {
-        padding: 0.5rem;
-        text-align: center;
-        background-color: red;
-        width: 100%;
-        border-radius: 5px;
-        font-family: var(--body-font);
-        font-size: 12px;
-        font-weight: 600;
-        cursor: pointer;
-        color: #ffffff;
-        border: none;
-        outline:none;
-        @media screen and (max-width: 576px) {
-          font-size: 0.75rem;
-          padding: 0.25rem;
-        }
-      }
-    }
-  }
-}
